Moved get_vivo_academic_articles from vivopump.py into pub_match_filter.py due to sourcing issues
